split sri wip
