+++ core-ui/src/components/Extensibility/plugins/SchemaRulesInjector.js
@@ -2,9 +2,6 @@ import React from 'react';
 import { last } from 'lodash';
 import { getNextPlugin } from '@ui-schema/ui-schema/PluginStack';
 import { List, fromJS } from 'immutable';
-import * as jp from 'jsonpath';
-
-import { useVariables } from '../hooks/useVariables';
 
 // fake an OrderedMap-like structure using List to allow for duplicate keys
 const propertiesWrapper = src => ({
@@ -13,40 +10,21 @@ const propertiesWrapper = src => ({
 
 export function SchemaRulesInjector({
   schema,
-  storeKeys,
+  // storeKeys,
   currentPluginIndex,
   rootRule,
   value,
   resource,
   ...props
 }) {
-  const { vars, setVar } = useVariables();
+  // console.log('SchemaRulesInjector', props.storeKeys.join('.'));
+  // console.log('props', props);
 
   const nextPluginIndex = currentPluginIndex + 1;
   const Plugin = getNextPlugin(nextPluginIndex, props.widgets);
 
-  const varName = schema.get('var');
   const { simple, advanced, path: myPath, children: childRules, ...itemRule } =
     schema.get('schemaRule') ?? rootRule;
-
-  if (varName) {
-    const varSuffix = storeKeys
-      .filter(item => typeof item === 'number')
-      .map(item => `[${item}]`)
-      .join('');
-    const varPath = `$.${varName}${varSuffix}`;
-
-    return (
-      <Plugin
-        {...props}
-        currentPluginIndex={nextPluginIndex}
-        schema={schema}
-        value={jp.value(vars, varPath)}
-        onChange={e => setVar(varPath, e.data.value)}
-        storeKeys={fromJS([schema.get('var')])}
-      />
-    );
-  }
 
   let newSchema = schema.mergeDeep(itemRule);
 
@@ -80,7 +58,6 @@ export function SchemaRulesInjector({
       {...props}
       currentPluginIndex={nextPluginIndex}
       schema={newSchema}
-      storeKeys={storeKeys}
       value={value}
       resource={resource}
     />
